Working on the addition operation
Buttons 0-9 output their corresponding values to the screen when pressed.
When '+' button is pressed, following a number, the the results window is updated.

However, executing '1 + 2' leads to value '2' in the results window. This is incorrect.

As well, '1 + 2 + 3' results in value '4'. This indicates that the intermediate storage value is one step behind the rest of the '=' button's operations.